setup spawnEdit
